quilt-steel-poc: Added support for simple ELSE bucket
Added a special state time series for recording mismatched time sums
in a quilt window.

The mismatched time value can be either positive or negative.

As a special case this can be used as an "ELSE" bucket for time
that is spent in a state that is not specified by any of the defined
quilt states.

Signed-off-by: Ulf Stroberg <argus-iot@strobergs.net>